requirements.txt
# Sistemsel ve yardımcı kütüphaneler
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pandas==2.2.3
numpy==2.0.2
tqdm==4.67.1
requests==2.32.3
matplotlib==3.10.0
seaborn==0.13.2

# Deep learning
torch==2.6.0
torchvision==0.21.0
timm==1.0.15
torchaudio==2.6.0
torchinfo==1.8.0
optuna==4.3.0

# CV ve görsel işlemler
opencv-python==4.10.0.84
Pillow==11.1.0
scikit-image==0.25.2
image-quality==1.2.7

# ML araçları
scikit-learn==1.6.1
PyWavelets==1.8.0
brisque==0.0.16
piq==0.8.0

# Streamlit & Web Arayüz
streamlit==1.45.1

# Diğer (streamlit ile ilişkili veya bağımlı)
altair==5.5.0
pydeck==0.9.1
watchdog==6.0.0
pyyaml==6.0.2
